qmail-1.03-maildir-uniq patch
Some operating systems quickly recycle PIDs, which can lead
to collisions between Maildir-style filenames, which must
be unique and non-repeatable within one second.

This patch is just a means of updating qmail-local to use
the format of the revised Maildir protocol, available at:

http://cr.yp.to/proto/maildir.html

It uses four unique identifiers:
* inode number of the file written to Maildir/tmp
* device number of the file written to Maildir/tmp
* time in microseconds
* the PID of the writing process

A Maildir-style filename would look like the following:

In Maildir/tmp:
  time.MmicrosecondsPpid.host
In Maildir/new:
  time.IinodeVdeviceMmicrosecondsPpid.host

Additionally, this patch further conforms to the revised
Maildir protocol by looking through the hostname for
instances of '/' and ':', replacing them with "057" and
"072", respectively, when writing it to disk.

Special thanks go to Matthias Andree for design and
sanity-checking.

  --Toby Betts <tmb2@po.cwru.edu>

http://vorlon.case.edu/~tmb2/qmail-1.03/qmail-1.03-maildir-uniq.patch

Signed-off-by: Giam Teck Choon <giamteckchoon@gmail.com>