qmail-smtpd-auth-close3
qmail-smtpd-auth-0.31 has a minor bug if you use morercpthosts. After a failed
authentication attempt, if the client attempts to send mail to a domain which
is not listed in rcpthosts, qmail-smtpd is unable to read morercpthosts.cdb.
Instead, it sends "421 unable to read controls (#4.3.0)" and drops the
connection. qmail-smtpd-auth-close3.patch fixes this bug by closing file
descriptor 3 (only if necessary) in the authentication child process rather
than the parent process.

Patch downloaded from:
  http://tomclegg.net/qmail/qmail-smtpd-auth-close3.patch

Signed-off-by: Giam Teck Choon <giamteckchoon@gmail.com>