qmail-1.03-domainbindings-1.2
NOTE: This is a modified patch to support IPv6 addresses as well hence the
separator character we use is '|' instead of ':' to support parsing of IPv6
addresses.

control/domainbindings for IPv4
control/domainbindings6 for IPv6

DomainBindings by Charles Cazabon

For servers that host multiple domains or have multiple IP addresses assigned
to them, it is sometimes useful (or important) to have qmail use a specific IP
address for its outgoing mail. By default, qmail uses whatever address the OS
chooses for all outbound connections. With this patch, you can specify which
address to use. It uses a control file similar to smtproutes to specify the
outbound IP address to use, based on the sender's domain.

Original patch:
  http://pyropus.ca/software/misc/qmail-1.03-domainbindings-1.2.patch

Signed-off-by: Giam Teck Choon <giamteckchoon@gmail.com>